+++ plugins/tracker-resources/src/components/issues/edit/CreateSubIssue.svelte
@@ -18,9 +18,8 @@
   import presentation, { getClient, KeyedAttribute } from '@hcengineering/presentation'
   import { StyledTextArea } from '@hcengineering/text-editor'
   import { IssueStatus, IssuePriority, Issue, Team, calcRank } from '@hcengineering/tracker'
-  import { Button, Component, EditBox } from '@hcengineering/ui'
+  import { addNotification, Button, Component, EditBox } from '@hcengineering/ui'
   import tags, { TagElement, TagReference } from '@hcengineering/tags'
-  import { addNotification } from '../../../utils'
   import tracker from '../../../plugin'
   import AssigneeEditor from '../AssigneeEditor.svelte'
   import StatusEditor from '../StatusEditor.svelte'
@@ -127,7 +126,7 @@
         })
       }
 
-      addNotification(generateId(), tracker.string.IssueCreated, getTitle(newIssue.title), IssueNotification, {
+      addNotification(tracker.string.IssueCreated, getTitle(newIssue.title), IssueNotification, {
         issueId: objectId,
         subTitlePostfix: (await translate(tracker.string.Created, { value: 1 })).toLowerCase()
       })
